+++ python/GafferRenderManTest/__init__.py
@@ -45,7 +45,7 @@ from .RenderManDisplayFilterTest import RenderManDisplayFilterTest
 from .RenderManSampleFilterTest import RenderManSampleFilterTest
 from .RenderManRenderTest import RenderManRenderTest, RenderManXPURenderTest
 from .InteractiveRenderManRenderTest import InteractiveRenderManRenderTest, InteractiveRenderManXPURenderTest
-from .RenderPassAdaptorTest import RenderPassAdaptorTest
+from .RenderPassAdaptorTest import RenderPassAdaptorTest, XPURenderPassAdaptorTest
 from .RenderManLightFilterTest import RenderManLightFilterTest
 from .StylizedAOVAdaptorTest import StylizedAOVAdaptorTest